+++ CHANGELOG.md
@@ -7,6 +7,7 @@ file.
 ### Added
 - Added Benchmarks to `RunType` to allow coverage of benchmark tests to be collected
 - Added Examples to `RunType` to allow coverage of examples to be collected
+- Instructions for integration to Gitlab pipelines to `README.md`
 
 ### Changed