+++ CHANGELOG.md
@@ -8,6 +8,7 @@ file.
 - Added Benchmarks to `RunType` to allow coverage of benchmark tests to be collected
 - Added Examples to `RunType` to allow coverage of examples to be collected
 - Instructions for integration to Gitlab pipelines to `README.md`
+- `--no-run` option to build tests and not collect coverage
 
 ### Changed